progetto_cielo/Rest_prog/requirements.txt
psycopg2-binary == 2.9.1
Flask>=2.0
Werkzeug>=2.0